backend/requirements.txt
# FastAPI and server
fastapi==0.120.4
uvicorn==0.38.0
python-multipart==0.0.20

# Database
neo4j==6.0.2
psycopg2-binary==2.9.11
sqlalchemy==2.0.44

# Authentication & Security
python-jose==3.5.0
passlib==1.7.4
pydantic==2.12.3
pydantic-settings==2.11.0
bcrypt==4.0.1

# AI Integration
openai==2.6.1

# HTTP Client
httpx==0.28.1

# Caching
redis==7.0.1

# Environment
python-dotenv==1.2.1

# Utilities
email-validator==2.3.0

# Other
requests==2.32.5
python-dateutil==2.9.0.post0
